+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - Added `Node` class to the PowerElectronics module to separate nodes from circuit components.
 - Refactored Jacobian assembly in `PowerElectronics` module to reuse the CSR pattern.
 - Refactored Jacobian assembly in `PhasorDyanmcics` module to reuse the CSR pattern.
+- Added LoadZIP model component type.
 ## v0.1
 
 - Refactored code to support adding different model families.